added semi-dynamic env variables setup in workflows
updated doc, changelog, version (3.1)
fixed env bug in local scheduler (any env passed in a job would erase all default env variables)

+++ python/soma_workflow/test/workflow_tests/workflow_test.py
@@ -194,7 +194,8 @@ class WorkflowTest(unittest.TestCase):
                 cls.setup_wf_controller(None) # del WorkflowController
                 if config.get_mode() == LIGHT_MODE:
                     if not kwargs.get('keep_temporary', False):
-                        os.unlink(config._database_file)
+                        if os.path.exists(config._database_file):
+                            os.unlink(config._database_file)
                         if os.path.exists(config._database_file + '-journal'):
                             os.unlink(config._database_file + '-journal')
                         shutil.rmtree(config._transfered_file_dir)
